requirements-dev.txt
# Development and Testing Dependencies
# Linear Regression Guide

# Testing framework and plugins
pytest>=7.4.0,<9.0.0              # Unit testing framework
pytest-cov>=4.1.0,<6.0.0          # Code coverage measurement
pytest-timeout>=2.1.0,<3.0.0      # Test timeout control
pytest-xdist>=3.3.0,<4.0.0        # Parallel test execution

# Code formatting and linting
black>=24.0.0,<25.0.0             # Opinionated code formatter
flake8>=7.0.0,<8.0.0              # Style guide enforcement
mypy>=1.8.0,<2.0.0                # Static type checking
pre-commit>=3.6.0,<4.0.0          # Git hooks for code quality

# Type stubs for better type checking
types-requests>=2.31.0,<3.0.0     # Type stubs for requests library
pandas-stubs>=2.0.0,<3.0.0        # Type stubs for pandas library

# Property-based testing
hypothesis>=6.80.0,<7.0.0          # Property-based testing framework

# Security scanning (for CI/CD and development)
safety>=3.0.0,<4.0.0              # Check for known security vulnerabilities
